capstone_project/capstone_django/capstone_project/requirement.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-freeze==0.6.3
djangorestframework==3.12.2
idna==2.10
psycopg2-binary==2.8.6
pytz==2021.1
requests==2.25.1
soupsieve==2.2
sqlparse==0.4.1
urllib3==1.26.3
xmltodict==0.12.0
